requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7
numpy==1.23.5
pandas==2.0.1
scikit-learn==1.2.2
tensorflow==2.12.0
xgboost==1.7.5
matplotlib==3.7.1
seaborn==0.12.2
boto3==1.26.130
python-multipart==0.0.6
pytest==7.3.1
httpx==0.24.1